refactor: move ledger plugins to their own folder
