+++ src/services/ledgers.js
@@ -1,7 +1,7 @@
 'use strict'
 const config = require('./config')
 const log = require('../common').log
-const Multiledger = require('../lib/ledgers/multiledger')
+const Multiledger = require('../lib/multiledger')
 
 module.exports = new Multiledger({
   config: config,
